src/agent/requirements.txt
redis==7.1.0
pandas>=2.3.3
pyarrow>=22.0.0
fastapi==0.127.0
uvicorn==0.40.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyspark==3.3.1
spark-nlp==6.2.3